pubsub/requirements.txt
google-cloud-pubsub==0.32.1
Flask==0.12.2
gunicorn==19.7.1
